refactor: extract tests to sibling files for 4 more god-files
Moves inline #[cfg(test)] mod tests blocks into sibling _tests.rs
files to bring production files under the 500-line guideline.
No behavior change.

  skills/ops.rs          2678 -> 1711 (-967)
  security/policy.rs     1761 ->  817 (-944)
  providers/reliable.rs  1679 ->  861 (-818)
  composio/ops.rs        1583 ->  901 (-682)

Wired via `#[cfg(test)] #[path = "{name}_tests.rs"] mod tests;`
Same pattern as prior ef411cec (channel.rs, compatible.rs).
